Beginner/weather_checker/weather.py
import requests
import os
from dotenv import load_dotenv

# Load API key from .env file
load_dotenv()
API_KEY = os.getenv("API_KEY")
BASE_URL = "http://api.openweathermap.org/data/2.5/weather"

def get_weather(city):
    params = {
        "q": city,
        "appid": API_KEY,
        "units": "metric"
    }

    response = requests.get(BASE_URL, params=params)

    if response.status_code == 200:
        data = response.json()
        temp = data["main"]["temp"]
        description = data["weather"][0]["description"]
        humidity = data["main"]["humidity"]

        print(f"\n🌍 Weather in {city.capitalize()}:")
        print(f"🌡️ Temperature: {temp}°C")
        print(f"💧 Humidity: {humidity}%")
        print(f"📋 Description: {description.capitalize()}")
    else:
        print("\n❌ City not found or error with API.")

# Run the program
if __name__ == "__main__":
    city_input = input("Enter city name: ")
    get_weather(city_input)
