Beginner/weather_checker/requirements.txt
requests
python-dotenv
# pip install -r requirements.txt
